requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blessed==1.20.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0   
Django==4.2.5
django-cors-headers==4.2.0
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
emoji==2.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
joblib==1.3.2
nltk==3.8.1
numpy==1.26.0
pandas==2.1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==3.5.3
regex==2023.10.3
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
zope.interface==6.0
